+++ modules/mauth-authenticator-akka-http/src/main/scala/com/mdsol/mauth/akka/http/MauthPublicKeyProvider.scala
@@ -43,9 +43,9 @@ class MauthPublicKeyProvider(configuration: AuthenticatorConfiguration, signer: 
     * @return { @link PublicKey} registered in MAuth for the application with given appUUID.
     */
   override def getPublicKey(appUUID: UUID): Future[Option[PublicKey]] =
-    getPublicKeyF(appUUID).unsafeToFuture()
+    getPublicKeyIO(appUUID).unsafeToFuture()
 
-  override def getPublicKeyF(appUUID: UUID): IO[Option[PublicKey]] = memoizeF(Some(configuration.getTimeToLive.seconds)) {
+  override def getPublicKeyIO(appUUID: UUID): IO[Option[PublicKey]] = memoizeF(Some(configuration.getTimeToLive.seconds)) {
     val signedRequest = signer.signRequest(UnsignedRequest.noBody("GET", new URI(configuration.getBaseUrl + getRequestUrlPath(appUUID)), headers = Map.empty))
     retrievePublicKey()(IO.fromFuture(IO(HttpClient.call(signedRequest.toAkkaHttpRequest))))
   }